# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CheerConnect API",
    "description": "REST API for CheerConnect — a social network for the cheerleading community. Built with Next.js 16 App Router, PostgreSQL via Prisma ORM, and NextAuth.js for authentication.",
    "version": "1.0.0",
    "contact": {
      "name": "CheerConnect"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth.js JWT session cookie. Obtain by logging in via POST /api/auth/callback/credentials or OAuth flow."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          }
        },
        "required": [
          "error"
        ]
      },
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "Position": {
        "type": "string",
        "enum": [
          "FLYER",
          "BASE",
          "BACKSPOT",
          "FRONTSPOT",
          "TUMBLER",
          "COACH",
          "CHOREOGRAPHER",
          "JUDGE",
          "OTHER"
        ]
      },
...
      "TeamCategory": {
        "type": "string",
        "enum": [
          "ALLSTAR",
          "SCHOOL",
          "COLLEGE",
          "RECREATIONAL",
          "PROFESSIONAL"
        ]
      },
...
      "TeamSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "logo": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "PostAuthor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "positions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Position"
            }
          }
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "videoUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "isEdited": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "author": {
            "$ref": "#/components/schemas/PostAuthor"
          },
          "team": {
            "$ref": "#/components/schemas/TeamSummary"
          },
          "originalPost": {
            "nullable": true,
            "description": "Present when this post is a repost",
            "allOf": [
              {
                "$ref": "#/components/schemas/Post"
              }
            ]
          },
          "isLiked": {
            "type": "boolean"
          },
          "_count": {
            "type": "object",
            "properties": {
              "likes": {
                "type": "integer"
              },
              "comments": {
                "type": "integer"
              },
              "reposts": {
                "type": "integer"
              }
            }
          }
        }
      },
...
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Não autorizado"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
...
  "paths": {
...
    "/api/teams/{slug}": {
      "get": {
        "tags": [
          "Teams"
        ],
        "summary": "Get team details",
        "description": "Returns full team details including members, recent posts, and upcoming events.",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Team details"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": [
          "Teams"
        ],
        "summary": "Update team info",
        "description": "Requires `hasPermission` on the team.",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  },
                  "location": {
                    "type": "string",
                    "nullable": true
                  },
                  "category": {
                    "$ref": "#/components/schemas/TeamCategory"
                  },
                  "level": {
                    "type": "string",
                    "nullable": true
                  },
                  "website": {
                    "type": "string",
                    "format": "uri",
                    "nullable": true
                  },
                  "instagram": {
                    "type": "string",
                    "nullable": true
                  },
                  "logo": {
                    "type": "string",
                    "nullable": true
                  },
                  "banner": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Team updated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Teams"
        ],
        "summary": "Delete team",
        "description": "Requires `isAdmin` on the team.",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Team deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/teams/{slug}/posts": {
      "post": {
        "tags": [
          "Teams"
        ],
        "summary": "Create a team post",
        "description": "Post on behalf of a team. Requires `hasPermission`.",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content"
                ],
                "properties": {
                  "content": {
                    "type": "string",
                    "minLength": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Team post created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "post": {
                      "$ref": "#/components/schemas/Post"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/teams/{slug}/invites/{id}": {
      "delete": {
        "tags": [
          "Teams"
        ],
        "summary": "Cancel a team invite",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Invite ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Invite cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/teams/{slug}/media": {
      "post": {
        "tags": [
          "teams"
        ],
        "summary": "POST /api/teams/{slug}/media",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  }
}